refactor: move non-vulnerability tests from vulnerabilities.spec.js to AuthenticationAdaptersV2.spec.js
